Handle clicks on descendants of links correctly in window event handler
